### 2026-07-15 #10969 @editorialbot
[PRE REVIEW]: RogueGPT: A Controlled Stimulus Generation Framework for AI-Generated Misinformation Research
**Submitting author:** @aloth (<a href="http://orcid.org/0009-0003-9327-6865">Alexander Loth</a>)
**Repository:** https://github.com/aloth/RogueGPT
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.0
