### 2026-07-18 9a8a143
Correct CRED-1 domain count to 2,673
