### 2026-07-18 a6690ad
Enrich README from Camera-Ready GI paper: comparison table, ABS-mapping, agentic framing, intended users, limitations
