+++ README.md
@@ -14,7 +14,11 @@
 
 ## Motivation
 
-Empirical research on human perception of AI-generated text requires stimuli that are diverse, reproducible, and generated under controlled experimental conditions. RogueGPT provides a systematic framework for producing such stimuli across multiple LLM families, languages, journalistic styles, and content formats.
+Empirical research on AI-generated news increasingly needs stimuli that are diverse, reproducible, and generated under controlled experimental conditions - not only for human-perception studies, but for the broader shift from static fake-news generation to **agentic disinformation campaigns**. RogueGPT provides a systematic, multi-model framework for producing such stimuli across multiple LLM families, languages, journalistic styles, and content formats, with full provenance tracking.
+
+Unlike single-model predecessors (e.g., GROVER), RogueGPT uses an extensible, model-agnostic architecture that connects static text generation with dynamic **agent-based simulation (ABS)** of misinformation diffusion. By standardizing the "seeding" phase of misinformation simulations, it supports red-teaming of detection systems against diverse, parameter-rich content.
+
+**Intended users.** The infrastructure targets four overlapping communities: (i) **misinformation researchers** who need controlled stimuli linked to perception data; (ii) **red-teamers** who generate adversarial content variants to stress-test detectors; (iii) **agent-based simulation researchers** who need reproducible seed content for diffusion models; and (iv) **human-factors researchers** studying perception.
 
 RogueGPT is the stimulus generation component of the [JudgeGPT](https://github.com/aloth/JudgeGPT) research pipeline. Together, they enable end-to-end experiments from controlled content generation to quantitative human perception measurement. This methodology is described in our foundational survey, *"Blessing or Curse? A Survey on the Impact of Generative AI on Fake News"* ([arXiv:2404.03021](https://arxiv.org/abs/2404.03021)), and extended in two WWW '26 Companion papers: *"Industrialized Deception"* ([arXiv:2601.21963](https://arxiv.org/abs/2601.21963)), which examines the systemic effects of LLM-generated misinformation, and *"Eroding the Truth-Default"* ([arXiv:2601.22871](https://arxiv.org/abs/2601.22871)), which reports the human perception findings from stimuli generated by this framework.
 
@@ -96,6 +100,46 @@ RogueGPT operates as the first stage of a two-part experimental workflow:
 2. **Storage (MongoDB):** Each fragment is stored with full provenance metadata, enabling reproducible filtering by any experimental variable.
 3. **Human evaluation ([JudgeGPT](https://github.com/aloth/JudgeGPT)):** Participants assess fragments on continuous dual-axis scales (source attribution and authenticity), producing quantitative perception data linked to generation parameters.
 4. **Analysis:** The combined dataset supports investigations into model-specific detectability, cross-linguistic perception differences, and the role of individual differences in judgment accuracy.
+
+## Comparison with Existing Solutions
+
+RogueGPT (generation) and JudgeGPT (evaluation) are positioned against notable fake-news generation systems along dimensions that matter for seeding agent-based simulations and red-teaming detectors. The comparison is *feature-scoped* (design intent and implemented functionality), not a claim of superior generation quality or detection performance.
+
+| Feature | **RogueGPT + JudgeGPT** | GROVER | FACTGEN | FakeGPT |
+|:---|:---:|:---:|:---:|:---:|
[…]
