### 2026-07-22 730bb21
Add SEO landing page: JSON-LD schema, canonical, sitemap.xml, robots.txt
