+++ docs/robots.txt
@@ -1,0 +1,4 @@
+User-agent: *
+Allow: /
+
+Sitemap: https://aloth.github.io/RogueGPT/sitemap.xml
