### 2026-07-29 609c71b
Fix conference month in citation metadata (WWW '26 held in June)
WWW '26 was originally scheduled for April 13-17, 2026, then
postponed to June 29 - July 3, 2026 and held virtually. Our
entries carried April and May, neither of which was correct.

Note that the ACM DL page for the main proceedings volume still
shows the original April dates in its header, so both dates
circulate. The Companion volume (ISBN 979-8-4007-2308-7), where
these papers appear, correctly lists June 29 - July 3. The
location field stays as Dubai, matching the ACM record.

Also replaced the stale "To appear" notes with t
[…]
