+++ CITATION.cff
@@ -85,7 +85,7 @@ preferred-citation:
       orcid: "https://orcid.org/0000-0001-5241-3809"
   collection-title: "Companion Proceedings of the ACM Web Conference 2026 (WWW '26 Companion)"
   year: 2026
-  month: 4
+  month: 6
   publisher:
     name: ACM
   location:
