### 2026-07-29 b146b72
Release 1.3.0: single version source, changelog, scope documentation
Version handling
- pyproject.toml is now the only place the version is declared. core.py
  resolves it at runtime via importlib.metadata, with a source-checkout
  fallback; app.py and mcp_server.py read core.__version__.
- This removes a four-way disagreement: pyproject.toml said 1.2.0,
  app.py and mcp_server.py said 1.1.0, and CITATION.cff said 0.1.0.
  The CFF value is what Zenodo and the GitHub cite button display.

Documentation
- Added CHANGELOG.md covering 1.0.0 through 1.3.0.
- Added a research-context section at the 
[…]
