### 2026-07-29 44f438a
Align corpus figures in paper.md with the README
paper.md still carried the pre-expansion numbers (>2,600 fragments,
222 human-authored) while the README reports the current corpus:
3,278 fragments, 2,638 machine-generated and 640 human-sourced.
The README figures are the correct ones.
