+++ paper/paper.md
@@ -46,13 +46,14 @@ normalisation layer, ensuring that every fragment — whether ingested manually,
 automated generation scripts, or through an AI agent — conforms to the same schema
 and that all experimental parameters are recorded for reproducibility.
 
-The current corpus contains more than 2,600 multilingual news fragments generated
-by 37 model configurations across 10 providers (OpenAI, Anthropic, Google, Meta,
-Mistral, DeepSeek, Microsoft, Zhipu, Moonshot, and Qwen), covering four languages
-(English, German, French, and Spanish), three content formats (tweet, headline, and
-short article), and five journalistic styles per language. Alongside the
-machine-generated material, 222 human-authored fragments — both legitimate and
-fake — serve as experimental anchors for perception studies.
+The current corpus contains 3,278 multilingual news fragments. Of these, 2,638 are
+machine-generated by 37 model configurations across 10 providers (OpenAI,
+Anthropic, Google, Meta, Mistral, DeepSeek, Microsoft, Zhipu, Moonshot, and Qwen),
+covering four languages (English, German, French, and Spanish), three content
+formats (tweet, headline, and short article), and five journalistic styles per
+language. The remaining 640 fragments are human-sourced — both legitimate news and
+authentic fake-news material — and serve as experimental anchors for perception
+studies.
 
 RogueGPT is the upstream stimulus generation component of a three-tool research
 pipeline: CRED-1 [@loth2025cred1] identifies unreliable news sources, RogueGPT
