### 2026-07-29 58e2727
ci: pin ruff to 0.15.22
The Lint job installed ruff unpinned, so every run picked up whatever
version was current on PyPI. ruff 0.16.0 (released between 22 and 29
July) enabled new rules; the same unchanged code that passed under
0.15.22 now reports 46 findings, mostly SIM117, UP006, UP045 and I001.

That is why CI went red on 609c71b, a commit that only touched
CITATION.cff and README.md and contained no Python at all.

Verified: commit 730bb21, the last green run, also reports 46 findings
under 0.16.0 and passes cleanly under 0.15.22. The regression is
entirely the toolchain, not the code.


[…]
