### 2026-07-29 50b6e43
Soften the ethics review wording in the impact statement
The paper and README stated that human data collection through JudgeGPT
"is conducted under" institutional ethics review, which reads as a claim
that a formal approval is already on file. The accurate statement is that
such collection is subject to institutional ethics review, matching the
existing wording in the README limitations section.

Precision matters here because the JOSS pre-review discussion
(openjournals/joss-reviews#10969) points reviewers at this section.
