### 2026-08-12 94a948e
docs: point Zenodo badge at concept DOI, add dataset version table
The badge and dataset links used the v1.0.0 version DOI (18703138), which
pins the README to a superseded 2,308-fragment release. Concept DOI 18703137
always resolves to the current version instead.

Adds a version table listing v1.1.0 (21904524, 3,278 fragments, 23 Mar 2026)
alongside the superseded v1.0.0, plus a note on duplicate content that leaks
across random train/test splits.
