+++ README.md
@@ -3,7 +3,7 @@
 [![CI](https://github.com/aloth/RogueGPT/actions/workflows/ci.yml/badge.svg)](https://github.com/aloth/RogueGPT/actions/workflows/ci.yml)
 [![arXiv](https://img.shields.io/badge/arXiv-2601.21963-B31B1B.svg)](https://arxiv.org/abs/2601.21963)
 [![arXiv](https://img.shields.io/badge/arXiv-2601.22871-B31B1B.svg)](https://arxiv.org/abs/2601.22871)
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.18703138.svg)](https://doi.org/10.5281/zenodo.18703138)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.18703137.svg)](https://doi.org/10.5281/zenodo.18703137)
 [![Status](https://img.shields.io/badge/status-active%20%26%20evolving-orange.svg)](https://github.com/aloth/RogueGPT)
 [![License](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
 [![GitHub Stars](https://img.shields.io/github/stars/aloth/RogueGPT?style=social)](https://github.com/aloth/RogueGPT/stargazers)
@@ -108,7 +108,18 @@ Of the full corpus, **2,638 fragments are machine-generated** and **640 are huma
 
 The corpus is available on Zenodo under restricted access for academic research:
 
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.18703138.svg)](https://doi.org/10.5281/zenodo.18703138)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.18703137.svg)](https://doi.org/10.5281/zenodo.18703137)
+
+### Dataset versions
+
+The concept DOI [`10.5281/zenodo.18703137`](https://doi.org/10.5281/zenodo.18703137) always resolves to the latest version. Cite a version DOI when your analysis has to be reproducible.
+
+| Version | Version DOI | Fragments | Snapshot | Status |
+|---|---|---|---|---|
+| 1.1.0 | [`10.5281/zenodo.21904524`](https://doi.org/10.5281/zenodo.21904524) | 3,278 | 23 Mar 2026 | Current |
+| 1.0.0 | [`10.5281/zenodo.18703138`](https://doi.org/10.5281/zenodo.18703138) | 2,308 | 19 Feb 2026 | Superseded |
+
+> **Note for machine learning use:** the corpus contains duplicate content under distinct fragment identifiers (545 rows across 196 repeated texts). Group by normalized text before creating train and test splits, otherwise a random split leaks training content into the test set.
 
 The model configuration (`prompt_engine.json`) currently defines **37 model identifiers** across 10 providers, enabling rapid expansion of the corpus with new model generations.
 
@@ -324,7 +335,7 @@ RogueGPT describes implemented functionality rather than a completed empirical e
 | Goal | Action |
 |:---|:---|
 | Understand the methodology | [Read the paper](https://arxiv.org/abs/2404.03021) |
-| Use the dataset | [Request access on Zenodo](https://doi.org/10.5281/zenodo.18703138) |
+| Use the dataset | [Request access on Zenodo](https://doi.org/10.5281/zenodo.18703137) (current version: [`21904524`](https://doi.org/10.5281/zenodo.21904524)) |
 | Extend the corpus | Fork, add models, submit a PR |
 | Participate in the study | [JudgeGPT Survey](https://judgegpt.streamlit.app/) |
 | Contact the authors | [alexander.loth@stud.fra-uas.de](mailto:alexander.loth@stud.fra-uas.de) |
