### 2026-08-16 e70d95a
Docs: report corpus composition and generator config separately (v1.3.1)
The README and paper described the released corpus using figures from
prompt_engine.json. The two evolve independently: the configuration is a
capability surface that grows as new models are added, while the corpus is a
fixed snapshot of what was actually generated and retained.

Corpus figures now come from roguegpt-fragments.csv (2026-03-23): 3,278
fragments (640 human, 2,638 machine) produced by 10 models across 6 providers,
in 4 languages, 3 formats and 5 journalistic styles per language. The Dataset
section now lists
[…]
