+++ paper/paper.md
@@ -47,8 +47,8 @@ automated generation scripts, or through an AI agent — conforms to the same sc
 and that all experimental parameters are recorded for reproducibility.
 
 The current corpus contains 3,278 multilingual news fragments. Of these, 2,638 are
-machine-generated by 37 model configurations across 10 providers (OpenAI,
-Anthropic, Google, Meta, Mistral, DeepSeek, Microsoft, Zhipu, Moonshot, and Qwen),
+machine-generated by 10 models across 6 providers (OpenAI, Google, Meta,
+Anthropic, Mistral, and Microsoft),
 covering four languages (English, German, French, and Spanish), three content
 formats (tweet, headline, and short article), and five journalistic styles per
 language. The remaining 640 fragments are human-sourced — both legitimate news and
@@ -124,8 +124,11 @@ available for academic research under restricted access.
 
 ## Multi-Model Support
 
-The model registry in `prompt_engine.json` currently defines 37 identifiers across
-10 providers. The naming convention (`provider_model-variant`) is provider-agnostic;
+The model registry in `prompt_engine.json` defines 47 identifiers across
+11 providers as of 2026-02-23, a capability surface that is deliberately broader
+than any single corpus snapshot: registry entries are available for generation and
+do not imply that a model contributed to the released corpus. The naming convention
+(`provider_model-variant`) is provider-agnostic;
 adding a new model requires a single line change to the JSON file and no code
 modifications. At ingest time, the CLI and MCP server validate submitted model
 identifiers against the registry, with an optional `--lenient` flag for
