+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "roguegpt"
-version = "1.3.0"
+version = "1.3.1"
 description = "A controlled stimulus generation framework for AI news authenticity research"
 readme = "README.md"
 license = { text = "GPL-3.0-or-later" }
