### 2026-08-25 128532c
Point the archive DOI at the current corpus version
CITATION.cff carried doi 10.5281/zenodo.18703138, which is corpus
version 1.0.0 from 2026-02-19. The README version table already marks
that release as superseded, so the citation metadata and the
documentation disagreed with each other. The current deposit is v1.1.0
under 10.5281/zenodo.21904524.

Also adds the Kill Chain paper (arXiv:2608.21389) to both the CITATION.cff
reference list and the README, since it analyses perception data collected
on stimuli generated with this framework and was not listed before.

The Blessing or Curse entry m
[…]
