### 2026-08-25 5b9103f
Complete the related publications and projects lists
The Related Publications list named two papers while the BibTeX section
above it carried four, so the prose showed a narrower picture of the work
than the citation blocks. Adds Industrialized Deception, Eroding the
Truth-Default, the Kill Chain paper (arXiv:2608.21389) and the Blessing or
Curse survey.

The Related Projects table omitted provenance-linkage, whose benchmark
audit uses this corpus as its in-domain material.
