### 2026-08-25 223ea5a
Point the DOI badge at the current corpus version
The badge linked 10.5281/zenodo.18703137, the concept DOI, while
CITATION.cff and the README citation block reference the v1.1.0 version
DOI 10.5281/zenodo.21904524. The badge now matches them, so a reader
clicking it lands on the same snapshot the citation metadata names.
