### 2026-08-27 #10969 @dosvk
@crvernon Yes, still interested -  happy to review
