### 2026-08-27 #10969 @prajwal-svm
> 👋 @prajwal-svm - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Thanks @crvernon! Absolutely, I'll block some time and review this one soon👌
