### 2026-08-27 #11219 @aloth
Thanks @crvernon for handling this, and @dosvk and @prajwal-svm for taking the time to review.

I am following this thread and will respond promptly to questions here, as well as to issues and pull requests opened against the RogueGPT repository. Happy to clarify anything about the setup or the design decisions as you work through the checklists.
