### 2026-08-28 aloth/RogueGPT#3 @dosvk
MCP server cannot be imported on a fresh install
**Checklist item**: Functionality / Installation

The MCP server can't be imported on a fresh install, on either the review version or main, though the failure mode differs.

At v1.2.0 the bound is `mcp[cli]>=1.0`, so `pip install ".[app,mcp]"` resolves mcp 2.1.1 today (`pip show mcp` confirms it's the official package), and the import fails with the SDK's own migration message:

```
ModuleNotFoundError: No module named 'mcp.server.fastmcp'. This is mcp 2.x, where FastMCP was renamed to MCPServer ... or pin 'mcp<2' to keep running v1 code.
```

On main the bound is `mcp[cli]>=1.0,<2`, which resolves 1.29.1 today, and that fails too:

```
mcp = FastMCP(
      ^^^^^^^^
TypeError: FastMCP.__init__() got an unexpected keyword argument 'version'
```

I dug into where the 1.x line broke. mcp==1.12.0 imports and starts fine; 1.13.0, 1.16.0, 1.20.0 and 1.29.1 all fail. It isn't just the `version` kwarg either: with `version=` removed, 1.13.0+ then rejects `description=`. Reducing the call to `FastMCP("RogueGPT")` imports cleanly on both 1.13.0 and 1.29.1, so the fix on the code side is dropping those two kwargs. Two smaller things I noticed while testing: v1.2.0 hard-codes `version="1.1.0"` in that call (main already reads it from core), and the `>=1.0` floor was never valid anyway, since `mcp.server.fastmcp` only exists from 1.2.0.

With the kwargs fixed and mcp==1.12.0 I also exercised the server over the protocol: `list_tools` returns `ingest_fragment` and `retrieve_fragments`, and a `retrieve_fragments` call returns real data, so the server itself is in good shape once it can start.

Suggested fix: drop the `version` and `description` kwargs, declare `mcp[cli]>=1.2,<2`, and add a smoke test that imports mcp_server (the failure is at import time, so an import is enough to catch it). One caveat there: CI currently installs `.[dev]`, which doesn't include the mcp extra, so the test env would need `.[dev,mcp]` for that test to run.

Environment: fresh venvs, macOS arm64, Python 3.12; reproduced via both `pip install ".[app,mcp]"` and requirements.txt.
