### 2026-08-28 aloth/RogueGPT#4 @dosvk
generate_batch.py exits 0 when every generation fails
**Checklist item**: Functionality (minor)

generate_batch.py always exits 0, even when every generation fails. Without API keys set I get:

```
Generating 3 fragments...

[google_gemini-3-flash] short article | en | Fox News
  Seed: European Union bans all social media for minors under 16...
  ERROR: GEMINI_API_KEY not set

[google_gemini-3-pro-preview] short article | en | CNN
  Seed: European Union bans all social media for minors under 16...
  ERROR: GEMINI_API_KEY not set

[openai_gpt-4.1] short article | en | CNN
  Seed: European Union bans all social media for minors under 16...
  ERROR: <urlopen error [Errno 8] nodename nor servname provided, or not known>

Done. Success: 0, Errors: 3
exit=0
```

For scripted or CI use it would help if the process returned nonzero when there are zero successes (or an `--strict` flag that does).

Purely a scripting-ergonomics point, the per-model error messages themselves are clear (e.g. "ERROR: GEMINI_API_KEY not set").
