### 2026-08-28 #11219 @dosvk
#### Progress update

Install and functionality testing done on macOS arm64, Python 3.12, with a local MongoDB in Docker. Unless stated otherwise, everything below was run against the v1.2.0 tag, since that's the version this issue lists. Two issues filed on the repo:

- MCP server can't be imported on a fresh install : https://github.com/aloth/RogueGPT/issues/3
- Generate_batch.py exits 0 when every generation fails: https://github.com/aloth/RogueGPT/issues/4

@crvernon one thing to settle before I finish the checklist. This issue lists v1.2.0, but v1.3.0 was tagged on 2026-07-29, a month before this review started, and its changelog says it added the Software design, Research impact statement and AI usage disclosure paper sections plus bounded dependencies. Should I review v1.3.0 (or main) rather than v1.2.0? I read the paper from main in any case, since that's what gets compiled.

What I can already confirm at v1.2.0: with a real MongoDB the full test suite passes 64/64, including the seven `requires_db` tests that CI deselects, and the README's CLI examples (stats, models, retrieve, ingest) all work verbatim. The repo hygiene is genuinely good: registered pytest markers with a documented deselect, CHANGELOG, CITATION.cff, CONTRIBUTING with a marker table, and CI across 3.10-3.12 with coverage.

#### On the paper

State of the field is one of the stronger ones I've come across. It engages GROVER, LIAR, TruthfulQA, Kreps et al., M4, MULTITuDE and FakeNewsNet with a clear three-point differentiation. The Research impact statement addresses the dual-use question through design rather than a disclaimer (no publishing or distribution path, researcher-supplied API credentials, ethics review on the human-data side). Both are better than typical.

Three things I'd still raise:

1. Reproducibility phrasing versus stored provenance:The schema records model identifier, prompt, language, style, format and timestamp, but sampling parameters aren't stored per fragment, and generate_batch.py hard-codes temperature 0.9 in three places. Since LLM APIs are non-deterministic anyway, "carries the metadata required to reproduce experimental conditions exactly" overpromises slightly. Either record the sampling parameters in the fragment (cheap, one field) or soften to "reproduce the experimental design" and add a one-line limitation noting exact regeneration isn't possible across provider updates.

2. All four cited uses are from the authors' own research programme. JOSS accepts that, but saying it more prominently would pre-empt the question rather than invite it.

3. The corpus Zenodo record (10.5281/zenodo.18703138) resolves but is access-restricted, with no publicly downloadable files. If that's a deliberate dual-use safeguard it's a defensible choice, but two things follow. The paper currently reads as if the corpus is openly archived, so a sentence stating the access model and the reason would make it accurate. And the JOSS checklist asks that original data be accessible to reviewers, so I'd need access (or an editor call that the restriction is acceptable) before I can tick that item. On the plus side, machine/human origin is embedded per record in the schema (Origin, MachineModel, MachinePrompt), which keeps the data attributable, and that fact deserves a sentence in the paper too, since it's the strongest answer to the dual-use reading of the project name. While clarifying the access model, a line on the corpus licence would help too, since the 640 human-sourced fragments are scraped news text and a data statement should cover their redistribution terms.

Checklist is filled to where these items allow; the version question above gates the rest.
