### 2026-08-28 45e0953
Fix MCP server import and batch exit code (#3, #4)
The MCP server could not be imported on a fresh install. FastMCP dropped the version and description keyword arguments in mcp 1.13; passing either raises TypeError on every release after 1.12. The call now passes only the server name.

The dependency floor was wrong in a second way. pyproject.toml declared mcp[cli]>=1.0 with no upper bound, so pip resolved mcp 2.x, where FastMCP has been renamed and the import fails with the SDK's migration message. requirements.txt already carried <2 but kept the same invalid floor: mcp.server.fastmcp does no
[…]
