+++ requirements.txt
@@ -9,4 +9,4 @@ pymongo>=4.0,<5
 dnspython>=2.0,<3      # required for mongodb+srv:// URIs
 openai>=1.0,<3
 streamlit>=1.30,<2
-mcp[cli]>=1.0,<2
+mcp[cli]>=1.2,<2
