+++ .github/workflows/ci.yml
@@ -28,7 +28,7 @@ jobs:
       - name: Install package + dev dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install ".[dev]"
+          pip install ".[dev,mcp]"
 
       - name: Run tests (skip DB-dependent tests)
         run: |
