### 2026-08-28 aloth/RogueGPT#3 @aloth
Thanks - this was precise enough to fix directly, and the version bisection saved me the work of finding where 1.x broke.

All three points confirmed and fixed:

- version= and description= dropped; FastMCP("RogueGPT") is all that remains
- mcp[cli]>=1.2,<2 in both pyproject.toml and requirements.txt

On the second point: the bounds were inconsistent between the two files. requirements.txt carried >=1.0,<2, but pyproject.toml still said >=1.0 with no upper bound, so pip install ".[mcp]" pulled mcp 2.x regardless of which branch you were on. That is why the failure mode differed between your two attempts. You are right that the >=1.0 floor was never valid either, since mcp.server.fastmcp only appears in 1.2.

Added tests/test_mcp_server.py, which imports the module and asserts the server name and both tool functions. It is guarded by importorskip so the default .[dev] run stays green, and CI now installs .[dev,mcp] as you noted it would have to.

Verified locally against mcp 1.29.1: import succeeds and both tools register.

Fixed in v1.3.1.
