### 2026-08-28 aloth/RogueGPT#4 @aloth
Agreed, fixed. generate_batch.py now exits 1 when no fragment was produced and at least one attempt failed.

I skipped the --strict flag: a non-zero exit on zero successes is the expected default rather than something to opt into. A dry run attempts nothing and still exits 0, so --dry-run stays usable as a check.

Fixed in v1.3.1.
