### 2026-08-28 #11219 @aloth
@dosvk Both issues you opened are fixed and released as v1.3.1:

- aloth/RogueGPT#3 - MCP server import: FastMCP kwargs removed, mcp[cli]>=1.2,<2 in both pyproject.toml and requirements.txt, import smoke test added, CI now installs the mcp extra.
- aloth/RogueGPT#4 - generate_batch.py now exits non-zero when nothing was generated.

Release: https://github.com/aloth/RogueGPT/releases/tag/v1.3.1

Archive DOI to follow once Zenodo has picked up the tag.
