### 2026-08-28 #11219 @dosvk
Thanks for quick response @aloth

**Verified on v1.3.1, both confirmed fixed**
Fresh venv, `pip install ".[app,mcp]"` now resolves mcp 1.29.1 and `import mcp_server` succeeds, and I can see the smoke test plus the `.[dev,mcp]` CI change in the tag. `generate_batch.py` with no keys gives `Success: 0, Errors: 3` and exits 1, while `--dry-run` still exits 0, which matches your reasoning on skipping `--strict`, the non-zero default is the right call. The bounds inconsistency between pyproject.toml and requirements.txt explaining the two different failure modes is a nice catch, that had puzzled me.

I've checked the Functionality checklist item now and will carry on reviewing against v1.3.1 unless @crvernon prefers otherwise.

One clarification so it doesn't get lost: The archive DOI you mention is the code archive for the release, while my remaining Data sharing item is about the corpus record (10.5281/zenodo.18703138), which is a separate, access-restricted deposit, just to make sure.
