### 2026-08-29 72a6ffd
Record sampling parameters per fragment, state the corpus access model, reference stable concept DOIs
Generation stored the model identifier and prompt but not the sampling parameters, while generate_batch.py hard-coded temperature and token limit in three separate provider calls. All three now read from a single SAMPLING constant and the values are written to the fragment as SamplingParams. validate_fragment reports a warning when the field is absent rather than defaulting it, so fragments ingested before this change stay identifiable instead of carrying a value nobody recorded.

The paper cl
[…]
