+++ paper/paper.bib
@@ -118,9 +118,9 @@
   title     = {{RogueGPT} Multilingual News Fragment Corpus},
   year      = {2026},
   publisher = {Zenodo},
-  doi       = {10.5281/zenodo.18703138},
-  url       = {https://doi.org/10.5281/zenodo.18703138},
-  note      = {Restricted access for academic research},
+  doi       = {10.5281/zenodo.18703137},
+  url       = {https://doi.org/10.5281/zenodo.18703137},
+  note      = {Concept DOI, resolves to the current version. Restricted access for academic research},
 }
 
 @article{kreps2022,
