### 2026-08-29 #11219 @aloth
@dosvk Thanks, these were the right things to push on. All three are addressed in 72a6ffd on main.

**Reproducibility phrasing.** You were right that it overpromised. `generate_batch.py` hard-coded temperature and token limit in three separate provider calls; they now read from one `SAMPLING` constant and the values are written to the fragment as `SamplingParams`. `validate_fragment` warns when the field is absent rather than defaulting it, so fragments ingested before this change stay identifiable instead of carrying a value nobody recorded. The paper now says the metadata reproduces the experimental design and states why exact regeneration is out of reach: the APIs are non-deterministic and providers revise models behind stable identifiers.

**Self-citation.** Stated plainly now, as you suggested: all four cited uses originate from our own research programme, which reflects the stage of adoption rather than external uptake.

**Corpus access and licence.** Two separate reasons, and the paper and README now say both. The 2,638 machine-generated fragments are CC BY 4.0. The 640 human-sourced fragments are excerpts of third-party news material that we cannot license onward, and they are shared under the research exception for text and data mining, which limits sharing to researchers rather than the general public. The second reason is the one you anticipated: a corpus of misinformation-like material should not be distributed without an accountable requester. Per-fragment provenance (`Origin`, `MachineModel`, `MachinePrompt`, source reference) keeps the two parts separable, and that is now in the paper too, since you are right that it is the strongest answer to the dual-use reading of the name.

Happy to grant you access directly for the review, just send a Zenodo access request on the record and I will approve it. Requests from researchers at academic or non-profit institutions are granted; the check is about the requester, not about the project.

One correction on the DOI: the record you looked at, 10.5281/zenodo.18703138, is v1.0.0 with 2,308 fragments. The current version is v1.1.0 with the 3,278 the paper describes. The paper was pointing at the version DOI, which is my mistake; it now uses the concept DOI 10.5281/zenodo.18703137, which always resolves to the current version. `archive_doi` had the same problem and now points at the software concept DOI.

On the version question, I am happy either way and will follow whatever @crvernon decides.
