### 2026-08-29 #11219 @dosvk
Noted @crvernon, thanks

@aloth thanks, the changes in 72a6ffd all check out on my side
SamplingParams is written per fragment now with the validation warning for older records, the paper states the design-level reproducibility claim with the limitation, the self-citation note is plain, and the licensing section covers both the CC BY part and the TDM exception for the human excerpts. The concept DOI also resolves to v1.1.0 for me, and the record metadata matches the paper's numbers (3,278 = 2,638 + 640). Nice work, and the duplicate-content note in the deposit description is a thoughtful touch.

One snag on the access request: the record doesn't offer me the option, logged in or not. Should you allow access requests or something ? Let me know if I am missing something here. I'll send mine as soon as the option appears, and that's the last item on my checklist.
