### 2026-08-29 #11219 @aloth
@dosvk That was on my side, not yours - access requests were switched off on the record, so there was nothing for you to find. They are enabled now, on both the current version and v1.0.0. Please try again when you get a chance.
