### 2026-08-29 #11219 @dosvk
Thanks @aloth, corpus verified.

Counts match the paper and codebook exactly (3,278 total, 2,638 machine / 640 human, ISO language totals to the row), the SHA-256 in the codebook matches the file, FragmentIDs are unique, and every machine fragment carries MachineModel and MachinePrompt. The codebook documenting the 1,021 empty Language/Format cells and the duplicate-content caveat is the kind of thing most deposits skip.

One small note: the deposit description mentions a per-record source reference, but the CSV has no source column for the human fragments, worth a codebook line on whether that's deliberate.

With that, Data sharing is ticked and my checklist is complete.

@crvernon from my side the submission is in good shape. And thanks @aloth for the pace throughout, same-day fixes and releases made this an easy review to close out.
