### 2026-08-29 1588766
Paper: state where the provenance of the human-sourced fragments actually lives
The licensing paragraph claimed every fragment records its origin via Origin, MachineModel, MachinePrompt and a source reference. There is no source column for the human-sourced rows in this CSV and there was never meant to be one, so the sentence sent a reviewer looking for a field that does not exist.

The text now says what is true: machine-generated rows carry MachineModel and MachinePrompt, and the human-sourced rows are traceable to outlet and URL through the linked JudgeGPT deposit, which holds HumanOutlet, 
[…]
