### 2026-08-29 #11219 @aloth
@dosvk Thanks for the close read, and you found a real inconsistency. The description was wrong, not the deposit: there is no source column for the human fragments in this CSV, and there was never meant to be one. Outlet and URL for the same fragments live in the linked JudgeGPT deposit (10.5281/zenodo.18703385), which the codebook covers under "Provenance of human-authored fragments" - my wording sent you looking for something in the wrong file.

Corrected in the paper in 1588766; the two record descriptions carry the same sentence and are being updated to match.
