### 2026-08-29 #11219 @dosvk
Makes sense, thanks @aloth , the JudgeGPT deposit is the right home for that and the updated wording reads clearly now.

That was my last open item. My checklist is complete and from my side I'm happy to recommend acceptance, @crvernon.

Genuinely nice work on this @aloth, the package is well engineered and the separation between core and the three interfaces held up well under testing.
