### 2026-08-31 #11219 @prajwal-svm
Hey @aloth, I sent a Zenodo access request for the RogueGPT [corpus](https://zenodo.org/records/21904524) for this review. Could you approve it when you have a chance? Thank you!
