### 2026-08-31 aloth/RogueGPT#5 @prajwal-svm
[Bug] v1.3.1 wheel omits `prompt_engine.json`
**What happened?**

While testing v1.3.1 for the [JOSS review](https://github.com/openjournals/joss-reviews/issues/11219), I installed RogueGPT in a fresh Python 3.11 virtual environment using:

```bash
python -m pip install ".[app,mcp]"
```

The installation completed successfully, and `which roguegpt` confirmed that the command came from the new virtual environment.

However, running `roguegpt models` failed because `prompt_engine.json` was missing from the installed package:

```text
FileNotFoundError: [Errno 2] No such file or directory:
'.../.venv-joss/lib/python3.11/site-packages/prompt_engine.json'
```

**Expected behavior:**

The project defines `roguegpt` as an installed console command and declares `prompt_engine.json` as package data. The wheel should include this file so the installed command can load the model configuration.

**Steps to reproduce:**
1. Check out `v1.3.1` and create a fresh Python 3.11 virtual environment.
2. Run `python -m pip install ".[app,mcp]"`.
3. Run `roguegpt models`.

**Environment:**

- OS: macOS arm64
- Version: RogueGPT v1.3.1 (`45e0953b`), Python 3.11.14
