### 2026-08-31 aloth/RogueGPT#6 @prajwal-svm
[Bug] `retrieve --is-fake false` is parsed as `true`
**What happened?**

While testing RogueGPT v1.3.1 for the [JOSS review](https://github.com/openjournals/joss-reviews/issues/11219), I ran:

```bash
roguegpt retrieve --is-fake false
```

The test database contained only records labelled `"IsFake": true`, and the command returned one of those records.

This appears to happen because [`--is-fake` uses `type=bool`](https://github.com/aloth/RogueGPT/blob/v1.3.1/cli.py#L92). Python treats any non-empty string, including `"false"`, as `True`.

**Expected behavior:**

`--is-fake false` should retrieve only records labelled `"IsFake": false`. `--is-fake true` should retrieve only records labelled `"IsFake": true`.

**Steps to reproduce:**

1. Configure `ROGUEGPT_MONGO_URI` for a database containing an `"IsFake": true` record.
2. Run `roguegpt retrieve --is-fake false`.
3. Observe that the returned record has `"IsFake": true`.

**Environment:**

- OS: macOS arm64
- Version: RogueGPT v1.3.1 (`45e0953b`), Python 3.11.14
