### 2026-09-01 aloth/RogueGPT#7 @prajwal-svm
[Bug] Empty "Choose Format" selection crashes the web app
**What happened?**

While testing RogueGPT v1.3.1 for the [JOSS review](https://github.com/openjournals/joss-reviews/issues/11219), I cleared every option from **Choose Format**.

The Streamlit app immediately crashed with:

```text
TypeError: replace() argument 2 must be str, not list
```

This occurs before clicking **Generate**, so no AI request or database write is needed to reproduce it. The empty list reaches [prompt.replace()](https://github.com/aloth/RogueGPT/blob/v1.3.1/app.py#L178-L182), which expects text.

**Expected behavior:**

The app should show a validation message or disable **Generate** when no format is selected. It should not crash.

**Steps to reproduce:**

1. Check out RogueGPT v1.3.1.
2. Run `python -m streamlit run app.py`.
3. Open the local web interface.
4. Clear every option from **Choose Format**.
5. Observe the error.

**Environment:**

- OS: macOS arm64
- Version: RogueGPT v1.3.1 (`45e0953b`), Python 3.11.14
