### 2026-09-01 aloth/RogueGPT#8 @prajwal-svm
[Paper/Docs] Clarify which generation parameters are recorded precisely
**What I found**

While reviewing RogueGPT v1.3.1 for the [JOSS review](https://github.com/openjournals/joss-reviews/issues/11219), I noticed that the paper and README make broad reproducibility claims:

- The [paper Summary](https://github.com/aloth/RogueGPT/blob/v1.3.1/paper/paper.md#L34-L47) says that "all experimental parameters are recorded."
- The [README](https://github.com/aloth/RogueGPT/blob/v1.3.1/README.md#L145-L150) says that all generation parameters are persisted.

However, some generation settings are not included in the saved record. For example, the batch generator uses [`temperature=0.9` and `max_tokens=1024`](https://github.com/aloth/RogueGPT/blob/v1.3.1/generate_batch.py#L101-L111), but the [saved document](https://github.com/aloth/RogueGPT/blob/v1.3.1/generate_batch.py#L189-L201) does not include those values.

The [web-generation path](https://github.com/aloth/RogueGPT/blob/v1.3.1/app.py#L209-L226) also saves the model, prompt, and selected content fields, but does not record sampling settings.

**Why this matters**

A researcher can identify the model and prompt, but cannot recover every generation setting from the stored record. Therefore, "all experimental parameters" and "all generation parameters" appear stronger than the current implementation guarantees.

**Suggested resolution**

Could the documentation and paper either:

1. Describe exactly which metadata fields are persisted by each generation interface, or
2. Persist the additional generation settings needed to support the broader reproducibility claims?
