### 2026-09-01 aloth/RogueGPT#9 @prajwal-svm
[Paper] Correct the MULTITuDE description and citation metadata
**What I found**

While reviewing RogueGPT v1.3.1 for the [JOSS review](https://github.com/openjournals/joss-reviews/issues/11219), I found two issues with the description and citation of MULTITuDE.

First, the [State of the Field section](https://github.com/aloth/RogueGPT/blob/v1.3.1/paper/paper.md#L210-L216) says:

> MULTITuDE covers 11 generators across 11 languages

The [original MULTITuDE publication](https://aclanthology.org/2023.emnlp-main.616/) describes **8 multilingual LLMs across 11 languages**.

Second, the [author list in `paper.bib`](https://github.com/aloth/RogueGPT/blob/v1.3.1/paper/paper.bib#L155-L169) does not match the official ACL record. The official authors are:

Dominik Macko, Robert Moro, Adaku Uchendu, Jason Lucas, Michiharu Yamashita, Matúš Pikuliak, Ivan Srba, Thai Le, Dongwon Lee, Jakub Simko, and Maria Bielikova.

**Suggested correction**

Could the paper:

1. Change "11 generators across 11 languages" to "8 multilingual LLMs across 11 languages"?
2. Replace the MULTITuDE bibliography author list with the official ACL citation metadata?
