### 2026-09-01 aloth/RogueGPT#10 @prajwal-svm
[Bug] Batch generation rejects models listed in `prompt_engine.json`
**What happened?**

While testing RogueGPT v1.3.1 for the [JOSS review](https://github.com/openjournals/joss-reviews/issues/11219), I used a model listed in `prompt_engine.json`:

```bash
python generate_batch.py --models openai_gpt-4o_2024-08-06 --n-seeds 1 --langs en --formats headline
```

The batch generator failed with:

```text
ERROR: 'openai_gpt-4o_2024-08-06'
Done. Success: 0, Errors: 1
```

The JSON registry contains this model, but [`generate_batch.py` has a separate three-model mapping](https://github.com/aloth/RogueGPT/blob/v1.3.1/generate_batch.py#L19-L32). The [router lookup](https://github.com/aloth/RogueGPT/blob/v1.3.1/generate_batch.py#L137-L152) therefore raises a `KeyError`.

This happens before any AI provider or MongoDB call, so no API credentials or running database are required to reproduce it.

**Expected behavior:**

A registered model should either be supported by batch generation, or the command should clearly report that the model is unsupported and document the batch-specific model list.

**Steps to reproduce:**

1. Check out RogueGPT v1.3.1.
2. Run the command above.
3. Observe the model-name error and zero successful generations.

**Environment:**

- OS: macOS arm64
- Version: RogueGPT v1.3.1 (`45e0953b`), Python 3.11.14
