### 2026-09-01 #11219 @prajwal-svm
## First-pass review checks

Hi @aloth,

I completed my first pass over RogueGPT v1.3.1 (`45e0953b`), covering installation, the CLI, the Streamlit interface, batch generation, documentation, and different sections of the paper.

I reproduced the following items locally and opened separate issues with the steps and expected behavior:

- https://github.com/aloth/RogueGPT/issues/5
- https://github.com/aloth/RogueGPT/issues/6
- https://github.com/aloth/RogueGPT/issues/7
- https://github.com/aloth/RogueGPT/issues/8
- https://github.com/aloth/RogueGPT/issues/9
- https://github.com/aloth/RogueGPT/issues/10

These are first-pass findings, not my final review. I will continue with the remaining paper and reference checks and retest the affected paths after fixes are available. We can track the technical details in the linked issues.

Thank you!
