### 2026-09-01 244e092
Fix packaging, boolean argument parsing, empty format selection and batch model routing
The wheel did not contain prompt_engine.json, so roguegpt models failed with FileNotFoundError on every installed copy and only worked from a source checkout. This is a flat layout with no package directory, so [tool.setuptools.package-data] had no package to attach the file to and setuptools dropped it without warning. It now ships as a data-file under share/roguegpt, and core.py resolves it through a search covering both a source checkout and an installed prefix. Verified by building the wheel and running
[…]
