+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "roguegpt"
-version = "1.3.1"
+version = "1.4.0"
 description = "A controlled stimulus generation framework for AI news authenticity research"
 readme = "README.md"
 license = { text = "GPL-3.0-or-later" }
@@ -60,15 +60,19 @@ Homepage       = "https://github.com/aloth/RogueGPT"
 Repository     = "https://github.com/aloth/RogueGPT"
 "Bug Tracker"  = "https://github.com/aloth/RogueGPT/issues"
 Documentation  = "https://github.com/aloth/RogueGPT#readme"
-DOI            = "https://doi.org/10.5281/zenodo.18703138"
+DOI            = "https://doi.org/10.5281/zenodo.20681920"
 
 # ── Flat-layout: include all top-level .py modules ──────────────────────────
+# prompt_engine.json is shipped via data-files rather than package-data: this is a
+# flat layout with no package directory, so there is no package for package-data to
+# attach the file to and it was silently dropped from the wheel. core.py resolves it
+# through importlib.resources with a source-checkout fallback.
 [tool.setuptools]
 py-modules = ["core", "cli", "mcp_server", "app", "generate_batch"]
+include-package-data = false
 
-# Include the config file in the package data
-[tool.setuptools.package-data]
-"*" = ["prompt_engine.json"]
+[tool.setuptools.data-files]
+"share/roguegpt" = ["prompt_engine.json"]
 
 # ── pytest ───────────────────────────────────────────────────────────────────
 [tool.pytest.ini_options]
