### 2026-09-01 b598fde
State which generation parameters are recorded, correct the MULTITuDE description
The paper claimed that all experimental parameters are recorded and the README that all generation parameters are persisted. Batch generation stores the model, the prompt and the sampling parameters; the web interface stores model and prompt but delegates sampling to provider defaults. Both documents now state which interface records what, rather than making a claim the implementation does not support. Reported in #8.

MULTITuDE was described as covering 11 generators across 11 languages. The publication reports 
[…]
