+++ paper/paper.bib
@@ -153,10 +153,10 @@
 }
 
 @inproceedings{macko2023multitude,
-  author    = {Macko, Dominik and Moro, Robert and Gagliardi, Adaku and
-               Pikuliak, Mat\'{u}\v{s} and Srba, Ivan and
-               Pov\'{a}\v{z}an\'{y}, Tom\'{a}\v{s} and
-               Simi\'{c}, Michal and Bielikova, Maria},
+  author    = {Macko, Dominik and Moro, Robert and Uchendu, Adaku and
+               Lucas, Jason and Yamashita, Michiharu and
+               Pikuliak, Mat\'{u}\v{s} and Srba, Ivan and Le, Thai and
+               Lee, Dongwon and Simko, Jakub and Bielikova, Maria},
   title     = {{MULTITuDE}: Large-Scale Multilingual Machine-Generated Text
                Detection Benchmark},
   booktitle = {Proceedings of the 2023 Conference on Empirical Methods in
