+++ paper/paper.md
@@ -43,8 +43,13 @@ logic (`core.py`) from user-facing interfaces: a Streamlit web application, a
 command-line interface (CLI), and a Model Context Protocol (MCP) server for
 AI-agent integration. All three interfaces share a single validation and
 normalisation layer, ensuring that every fragment — whether ingested manually, via
-automated generation scripts, or through an AI agent — conforms to the same schema
-and that all experimental parameters are recorded for reproducibility.
+automated generation scripts, or through an AI agent — conforms to the same schema.
+Each machine-generated fragment records the model identifier, the full prompt, the
+sampling parameters used by the batch generator, and the language, format and seed
+phrase where the interface supplies them. Fragments created through the web
+interface record model and prompt but not sampling settings, because that path
+delegates them to the provider defaults; the stored record therefore states what
+was fixed rather than implying that every generation setting is recoverable.
 
 The current corpus contains 3,278 multilingual news fragments. Of these, 2,638 are
 machine-generated by 10 models across 6 providers (OpenAI, Google, Meta,
@@ -131,10 +136,11 @@ to researchers rather than the general public. The restriction also keeps a corp
 of misinformation-like material from being distributed without an accountable
 requester. Every fragment records its own origin. The machine-generated rows carry
 `MachineModel` and `MachinePrompt`; the human-sourced rows are traceable to
-outlet and URL through the linked JudgeGPT deposit
-([10.5281/zenodo.18703385](https://doi.org/10.5281/zenodo.18703385)), which holds
-`HumanOutlet`, `HumanURL` and `IngestedVia` for the same fragments. The `Origin`
-column keeps the two parts separable and independently filterable.
+outlet and URL through the `HumanOutlet` and `HumanURL` columns, which the deposit
+carries alongside `IngestedVia` for every fragment. The companion JudgeGPT deposit
+([10.5281/zenodo.22226580](https://doi.org/10.5281/zenodo.22226580)) ships a
+byte-identical copy of the same corpus. The `Origin` column keeps the two parts
+separable and independently filterable.
 
 # Key Features
 
@@ -232,7 +238,7 @@ work uses a single model without a reusable generation framework.
 More recent efforts such as the M4 corpus [@wang2024m4] collect
 multi-generator outputs across domains and languages but distribute them as static
 datasets without tooling for reproducible expansion. Similarly, MULTITuDE
-[@macko2023multitude] covers 11 generators across 11 languages but is frozen at
+[@macko2023multitude] covers 8 multilingual LLMs across 11 languages but is frozen at
 its release configuration. FakeNewsNet [@shu2020fakenewsnet] provides a rich
 repository combining news content with social context but focuses on detection
 research with pre-collected articles rather than controlled stimulus generation.
