### 2026-09-01 8ce1d91
Release 1.4.0
Tags the two commits that followed v1.3.1 and were never released, together with the fixes for issues #5 to #10.

Dataset references point at the current releases: corpus 10.5281/zenodo.22225536 (v1.2.0) and perception data 10.5281/zenodo.22226580 (v1.2.0). The corpus release publishes the full 15-column schema, which is what makes the provenance statement in the paper accurate.

CITATION.cff cited the corpus DOI while describing the software. It now carries the software concept DOI 10.5281/zenodo.20681920, as does the DOI project URL in pyproject.toml.
