+++ CITATION.cff
@@ -2,11 +2,11 @@ cff-version: 1.2.0
 title: "RogueGPT: A Controlled Stimulus Generation Framework for News Authenticity Research"
 message: "If you use this software, please cite it as below."
 type: software
-version: "1.3.1"
-date-released: "2026-07-29"
+version: "1.4.0"
+date-released: "2026-09-01"
 license: GPL-3.0-or-later
 repository-code: "https://github.com/aloth/RogueGPT"
-doi: "10.5281/zenodo.21904524"
+doi: "10.5281/zenodo.20681920"
 url: "https://github.com/aloth/RogueGPT"
 abstract: >
   RogueGPT is a controlled stimulus generation framework for AI news authenticity
