### 2026-09-01 aloth/RogueGPT#5 @aloth
Fixed in v1.4.0.

Your diagnosis was right and the cause is worse than a missing declaration. This is a flat layout with no package directory, so `[tool.setuptools.package-data]` has no package to attach `prompt_engine.json` to, and setuptools drops it without warning. The declaration was present and never had any effect, which means no published install has ever worked. It only ran from a source checkout, which is how it escaped notice.

`prompt_engine.json` now installs as a data-file under `share/roguegpt`, and `core.py` resolves it through a search covering a source checkout, `sys.prefix` and `sys.base_prefix`, with an error that lists the paths it tried.

Verified the way you found it: built the wheel, confirmed the file is in it, installed into a fresh Python virtual environment and ran `roguegpt models`, which lists all 47 identifiers.
