### 2026-09-01 aloth/RogueGPT#6 @aloth
Fixed in v1.4.0.

`--is-fake` now uses a dedicated parser that accepts true/false, yes/no and 1/0 case-insensitively and raises `ArgumentTypeError` on anything else, so a typo fails loudly instead of silently selecting the true set.

Added regression tests, including one that asserts the parsed `False` actually reaches `core.get_random_fragments` rather than only being parsed correctly.
