### 2026-09-01 aloth/RogueGPT#7 @aloth
Fixed in v1.4.0.

Worth recording why the existing guard did not catch this. The line read `selections[0] if isinstance(selections, list) and selections else selections`, which looks defensive but returns the empty list on the else branch, because an empty list is falsy. The guard routed the exact input it was meant to prevent straight into `prompt.replace()`.

Empty selections now resolve to an empty string, and the interface names the placeholders that are still unfilled instead of failing on the first one.
