### 2026-09-01 aloth/RogueGPT#8 @aloth
Fixed in v1.4.0.

You are right that the claim was broader than the implementation. Part of it had already been narrowed before your report: batch generation records `SamplingParams` per fragment. The web-generation path still did not, and the wording in both the paper summary and the README was unchanged, so the overclaim stood.

Rather than persisting sampling settings on the web path, which delegates them to provider defaults and would mean recording values the interface does not control, both documents now state which interface records what. Batch generation records model, prompt and sampling parameters; the web interface records model and prompt.
