### 2026-09-01 aloth/RogueGPT#9 @aloth
Fixed in v1.4.0.

Both points confirmed against the source rather than taken over directly. The ACL Anthology abstract states 74,081 texts in 11 languages generated by 8 multilingual LLMs, and Crossref returns the same eleven authors you listed for 10.18653/v1/2023.emnlp-main.616.

The text now reads "8 multilingual LLMs across 11 languages" and the bibliography entry carries the official author list.
