### 2026-09-01 aloth/RogueGPT#10 @aloth
Fixed in v1.4.0, though not by merging the two lists.

`prompt_engine.json` registers 47 identifiers and `generate_batch.py` wires 3. That gap is intentional: the registry names every identifier the corpus may legitimately contain, including fragments generated through the web interface or by providers no longer wired in, while the batch script can only drive models with a provider call behind them. Collapsing them would either break corpus validation or promise batch support that does not exist.

What was wrong is the failure mode. A registered identifier produced a bare `KeyError`, indistinguishable from a typo. The router now separates the two cases: registered but not driveable by batch generation, and unknown, each with the list of supported models.
