### 2026-09-01 #11219 @aloth
Hi @prajwal-svm and @dosvk,

v1.4.0 is out: https://github.com/aloth/RogueGPT/releases/tag/v1.4.0

It carries two sets of changes.

@prajwal-svm, all six first-pass issues are fixed. Each issue has the specific finding, so only what is worth flagging here:

**#5 was broader than the reported case.** The declaration meant to ship `prompt_engine.json` never had any effect, because a flat layout provides no package for `package-data` to attach to. No published install has ever loaded the model registry; it only worked from a source checkout. Verified by building the wheel and running `roguegpt models` in a fresh virtual environment.

**#7 had a guard that looked correct.** It returned the empty list on the branch meant to prevent exactly that input, since an empty list is falsy.

**#10 was resolved without merging the two model lists.** The registry and the batch script cover different sets by design, so the fix is a specific error message rather than a unified registry. The reasoning is in the issue.

**#9** was checked against Crossref and the ACL Anthology record rather than applied directly; both points hold.

@dosvk, the changes you reviewed on 29 August were committed at the time but never tagged, so they ship here as part of 1.4.0: per-fragment `SamplingParams`, the explicit access model, the concept DOIs, and the provenance wording. Nothing in them changed since you checked them. The tag you verified against is now superseded by 1.4.0.

The dataset references also move to the current corpus release, 10.5281/zenodo.22225536, which publishes the full 15-column schema. That resolves the point you raised about the missing source reference from the other direction: the corpus deposit now carries `HumanOutlet`, `HumanURL` and `IngestedVia` itself, so the paper no longer has to point at the JudgeGPT deposit for it.

Test suite is at 61 passing, up from 57. The seven failures in `tests/test_core.py` are pre-existing and need a live MongoDB instance via `ROGUEGPT_MONGO_URI`.

Ready for your retest whenever it suits you.
