### 2026-09-02 #11219 @dosvk
Retested against the v1.4.0 tag, @aloth.

Everything holds: the wheel-installed package now loads the model registry , my earlier items all ship in the tag, and the suite runs 68 passed with a live MongoDB.

The corpus consolidation also closes out the source-reference note from my side: the concept DOI resolves to the v1.2.0 deposit and its schema carries HumanOutlet, HumanURL and IngestedVia directly.

Nothing changes on my checklist, it was already complete, and my recommendation of acceptance stands.
