### 2026-09-05 #11219 @prajwal-svm
Hi @aloth, thank you for the updates. I'm almost done with my review. I have two final clarification questions:

1. Could you please check whether the current disclosure fully follows the [JOSS AI policy](https://joss.readthedocs.io/en/latest/policies.html#author-ai-usage)? My understanding is that the disclosure should identify the AI systems and versions used and state where they were used, such as in the code, documentation, or manuscript. Could you confirm whether these details are missing or should be added/not?
2. Could you please confirm that the following details are correct?
    - The M4 author list in [paper.bib](https://github.com/aloth/RogueGPT/blob/v1.4.0/paper/paper.bib#L139-L152), is different compared with the [official publication](https://aclanthology.org/2024.eacl-long.83/).
    - The statement that *Eroding the Truth-Default* included 320 participants, as the [cited paper](https://arxiv.org/html/2601.22871) appears to report 154 participants.
    - The dataset author roles in `paper.bib`, compared with the creator and contributor roles shown on Zenodo.

Once I have clarity on these points, I’ll complete and post my final review. Thank you for your patience!
