### 2026-09-05 ec6782c
Correct bibliography metadata and complete the AI usage disclosure
The M4 entry listed seven authors, five of whom do not appear on the publication, while thirteen of the actual authors were missing. DOI, pages, venue and month were all correct. That combination identifies the cause: the bibliography was generated from a Zotero export by an AI assistant, and for this entry the author list was reconstructed rather than transcribed. It was never checked against the source record.

Because a fabricated author list is not an error to fix in isolation, every entry in paper.bib has been verified aga
[…]
