+++ paper/paper.md
@@ -61,7 +61,7 @@ authentic fake-news material — and serve as experimental anchors for perceptio
 studies.
 
 RogueGPT is the upstream stimulus generation component of a three-tool research
-pipeline: CRED-1 [@loth2025cred1] identifies unreliable news sources, RogueGPT
+pipeline: CRED-1 [@loth2026cred1data] identifies unreliable news sources, RogueGPT
 generates controlled stimuli, and JudgeGPT delivers them to human participants
 for perception measurement.
 
@@ -101,7 +101,7 @@ RogueGPT addresses this gap in three ways:
 
 The framework is part of a broader open-source research ecosystem for studying
 online misinformation. Upstream, the CRED-1 Domain Credibility Dataset
-[@loth2025cred1] provides the automated source selection that identifies
+[@loth2026cred1data] provides the automated source selection that identifies
 unreliable domains whose content is scraped as human-authored fake news anchors.
 Downstream, [JudgeGPT](https://github.com/aloth/JudgeGPT) delivers
 RogueGPT-generated stimuli to human participants and records dual-axis perception
@@ -112,16 +112,16 @@ RogueGPT is already in active use in the following peer-reviewed publications. A
 four originate from the authors' own research programme, which reflects the stage
 of adoption rather than a claim of external uptake:
 
-- *Industrialized Deception* [@loth2026industrialized] analyses the systemic
+- *Industrialized Deception* [@loth2026collateraleffects] analyses the systemic
   effects of LLM-generated misinformation on digital ecosystems, using stimuli
   generated by RogueGPT (ACM WWW '26).
 - *Eroding the Truth-Default* [@loth2026eroding] reports human perception data
-  from 320 participants who evaluated stimuli produced by this framework
-  (ACM WWW '26).
+  from 154 participants who contributed 918 evaluations of stimuli produced by
+  this framework (ACM WWW '26).
 - *The Verification Crisis* [@loth2026verification] presents expert survey
   findings on GenAI disinformation and reproducible provenance as part of the
   same research programme (ACM WWW '26).
-- The foundational survey *Blessing or Curse?* [@loth2024survey] describes the
+- The foundational survey *Blessing or Curse?* [@loth2024blessing] describes the
   methodology underpinning the generation pipeline.
 
 The full corpus produced by RogueGPT is archived on Zenodo
@@ -187,7 +187,7 @@ Human-written fragments are sourced from two pools:
 - **Legitimate news:** Opening paragraphs from established international outlets
   (BBC, Guardian, Reuters, Tagesschau, FAZ, Le Monde, El País) via RSS feeds.
 - **Fake news:** Articles scraped from domains classified as *fake*, *unreliable*,
-  or *conspiracy* by the CRED-1 Domain Credibility Dataset [@loth2025cred1],
+  or *conspiracy* by the CRED-1 Domain Credibility Dataset [@loth2026cred1data],
   which scores 2,672 domains using editorial classification, fact-check claims,
   web popularity, and domain age.
 
@@ -230,7 +230,7 @@ to a single GPT-2-scale architecture and English text. The LIAR dataset
 [@wang2017liar] provides 12,836 human-labelled claims across six fine-grained
 labels but offers no generation infrastructure. TruthfulQA [@lin2022truthfulqa]
 measures model truthfulness across 817 questions but targets factual consistency
-rather than controlled stimulus generation. Kreps et al. [@kreps2022] demonstrate
+rather than controlled stimulus generation. @kreps2022 demonstrate
 that GPT-2-generated political news is perceived as credible as human-written
 content, establishing the need for systematic multi-model comparisons — but their
 work uses a single model without a reusable generation framework.
[…]
