+++ paper/paper.bib
@@ -1,4 +1,4 @@
-@inproceedings{loth2026industrialized,
+@inproceedings{loth2026collateraleffects,
   author    = {Loth, Alexander and Kappes, Martin and Pahl, Marc-Oliver},
   title     = {Industrialized Deception: The Collateral Effects of {LLM}-Generated
                Misinformation on Digital Ecosystems},
@@ -8,6 +8,7 @@
   month     = jun,
   publisher = {ACM},
   address   = {New York, NY, USA},
+  pages     = {295--302},
   location  = {Dubai, United Arab Emirates},
   doi       = {10.1145/3774905.3795471},
   url       = {https://arxiv.org/abs/2601.21963},
@@ -21,7 +22,7 @@
                ({WWW} '26 Companion)},
   year      = {2026},
   month     = jun,
-  pages     = {1125--1133},
+  pages     = {1125--1132},
   publisher = {ACM},
   address   = {New York, NY, USA},
   location  = {Dubai, United Arab Emirates},
@@ -29,22 +30,27 @@
   url       = {https://arxiv.org/abs/2601.22871},
 }
 
-@article{loth2024survey,
-  author  = {Loth, Alexander and Kappes, Martin and Pahl, Marc-Oliver},
-  title   = {Blessing or Curse? {A} Survey on the Impact of Generative {AI}
-             on Fake News},
-  journal = {arXiv preprint arXiv:2404.03021},
-  year    = {2024},
-  url     = {https://arxiv.org/abs/2404.03021},
+@misc{loth2024blessing,
+  author        = {Loth, Alexander and Kappes, Martin and Pahl, Marc-Oliver},
+  title         = {Blessing or Curse? {A} Survey on the Impact of Generative {AI}
+                   on Fake News},
+  year          = {2024},
+  publisher     = {arXiv},
+  doi           = {10.48550/arXiv.2404.03021},
+  eprint        = {2404.03021},
+  archivePrefix = {arXiv},
+  primaryClass  = {cs.CL},
 }
 
-@dataset{loth2025cred1,
-  author    = {Loth, Alexander and Kappes, Martin and Pahl, Marc-Oliver},
-  title     = {{CRED-1}: Domain Credibility Dataset},
-  year      = {2025},
+@dataset{loth2026cred1data,
+  author    = {Loth, Alexander},
+  title     = {{CRED-1}: An Open Multi-Signal Domain Credibility Dataset},
+  year      = {2026},
   publisher = {Zenodo},
+  version   = {v1.0},
   doi       = {10.5281/zenodo.18769460},
   url       = {https://doi.org/10.5281/zenodo.18769460},
+  note      = {Version 1.0, the 2,672-domain snapshot used for source selection here. Concept DOI 10.5281/zenodo.18769459 always resolves to the latest version},
 }
 
 @inproceedings{zellers2019grover,
@@ -97,7 +103,7 @@
   url       = {https://arxiv.org/abs/2602.02100},
 }
 
-@inproceedings{loth2026originlens,
+@inproceedings{loth2026originlenswebsci,
   author    = {Loth, Alexander and Conceicao Rosario, Dominique and
                Ebinger, Peter and Kappes, Martin and Pahl, Marc-Oliver},
   title     = {Origin Lens: Reclaiming Trust on the {AI}-Mediated Web Through
@@ -113,9 +119,10 @@
   url       = {https://arxiv.org/abs/2602.03423},
 }
 
-@dataset{loth2026roguegpt_corpus,
-  author    = {Loth, Alexander and Kappes, Martin and Pahl, Marc-Oliver},
-  title     = {{RogueGPT} Multilingual News Fragment Corpus},
+@dataset{loth2026roguegptdata,
[…]
