### 2026-09-05 #11219 @aloth
Hi @prajwal-svm,

Thank you for both questions. All three points in your second question are correct, and the disclosure was indeed incomplete.

## 1. AI usage disclosure

You are right that it did not meet the policy: it named neither the system nor the version, and it described only the software side. The revised section names Claude Opus 4.8 used through GitHub Copilot in Visual Studio Code, separates where it was applied in the software (code completion, refactoring suggestions, docstrings) from where it was applied in the manuscript (generating the BibTeX bibliography from my Zotero export), and states explicitly that no AI system was used to design the study, to build or analyse the corpus, or to formulate the findings. The human verification the policy asks for is now affirmed in the same section.

## 2. The M4 author list

Confirmed, and it has the same origin as the disclosure gap. The entry listed seven authors, five of whom (Manber, Shang, Bensaid, He, Sharaf) do not appear on the publication, while thirteen of the actual authors were missing. DOI, pages, venue and month were all correct. That pattern is what identifies the cause: the record was real, but for this entry the author list was reconstructed instead of transcribed, and it was not verified against the ACL Anthology record at the time. The entry now carries the full fifteen-author list and the conference address.

A fabricated author list is not an error worth fixing in isolation, so I verified every entry in paper.bib against Crossref or DataCite rather than this one alone. That check found three further problems, listed under point 5.

## 3. The participant count

Confirmed. The cited paper reports 154 participants and 918 evaluations. The figure 320 appears nowhere in it. The sentence now states both numbers.

## 4. Dataset author roles

Confirmed. Both dataset entries listed three authors. The Zenodo records name only the depositor as creator, with the other two registered as contributors of type Supervisor. Both entries now match their deposits, and both titles have been corrected to the registered titles.

## 5. Further corrections from the full check

- Eroding the Truth-Default: pages 1125-1133 corrected to 1125-1132.
- CRED-1: the entry lacked its full title, the correct year and a version. The DOI is unchanged and deliberately so. 10.5281/zenodo.18769460 is version 1.0, the 2,672-domain snapshot the paper describes; the concept DOI resolves to a later version with a different domain count, which would no longer match the text.
- Self-citation keys now follow the canonical BibTeX source used across the research programme, so the same work carries the same key in every manuscript. This also separates the two Origin Lens publications, which previously shared a single key.
- Kreps et al. was cited with the author name written out before the citation, so the name appeared twice in the rendered text. It now uses a narrative citation.

All changes are in ec6782cdd3fb520fadb66e87cd29cf6c71f3c640. Thank you for catching this.
